Adjust object update priorities such that root prims are sent to the viewer before child prims.
This was originally a fix for huds, since child prims fail to display if the viewer doesn't receive the root prim update first.
However, on the advice of jhurliman, this has been done for all objects, both ordinary and attachments.
The separate mechanism in LLClientView which prevents child prim updates being sent out first is still present temporarily.
This is a foreport of the equivalent fix in 0.6.9-post-fixes, though that was for attachments only.