Protobuf all the things! Rewrite how sims are serialized and saved down into the .abackup file, and instead of using json and xml together, switch entirely to protobuf, which speeds up the loading and saving of sims. Clean up a lot of old nasty code with saving .abackup files for the region.
NOTE: This commit will migrate all of the sims that you start, and will spam the console with errors about changed properties (if the values it spits out are not very close together, SEND ME THE ERRORS). This will add new .sim files which include all of the region data in your bin folder, which replaces the old .abackup files.